api/requirements.txt
fastapi==0.115.6
pydantic>=2.0.0
PyYAML>=6.0
